Synced files to match parent dir tree
